# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TerrAPI is a service designed to simplify and automate your deployments.",
        "title": "TerrAPI Controller",
        "contact": {
            "name": "Support",
            "url": "https://github.com/terrapi-solution",
            "email": "contact@thomas-illiet.fr"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/deployments": {
          "get": {
            "security": [
              {
                "ApiKeyAuth": []
              },
              {
                "OAuth2Application": [
                  "write",
                  "admin"
                ]
              }
            ],
            "consumes": [
              "application/json"
            ],
            "produces": [
              "application/json"
            ],
            "tags": [
              "🍑 Deployment"
            ],
            "summary": "List all deployments.",
            "parameters": [
              {
                "type": "string",
                "description": "Search",
                "name": "search",
                "in": "query"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "collectionFormat": "csv",
                "description": "Filter",
                "name": "filter",
                "in": "query"
              },
              {
                "minimum": 1,
                "type": "integer",
                "default": 1,
                "description": "Page",
                "name": "page",
                "in": "query"
              },
              {
                "maximum": 100,
                "minimum": 1,
                "type": "integer",
                "default": 10,
                "description": "Page size",
                "name": "page_size",
                "in": "query"
              },
              {
                "type": "string",
                "default": "id",
                "description": "Order by",
                "name": "order_by",
                "in": "query"
              },
              {
                "type": "string",
                "default": "desc",
                "description": "Order direction",
                "name": "order_direction",
                "in": "query"
              }
            ],
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/definitions/deployments.ListResponse"
                }
              },
              "500": {
                "description": "Internal Server Error",
                "schema": {
                  "$ref": "#/definitions/errors.AppError"
                }
              }
            }
          },
          "post": {
            "consumes": [
              "application/json"
            ],
            "produces": [
              "application/json"
            ],
            "tags": [
              "🍑 Deployment"
            ],
            "summary": "Create a new deployment.",
            "parameters": [
              {
                "description": "Deployment",
                "name": "deployment",
                "in": "body",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/deployments.DeploymentRequest"
                }
              }
            ],
            "responses": {
              "201": {
                "description": "Created",
                "schema": {
                  "$ref": "#/definitions/deployments.DeploymentResponse"
                }
              },
              "500": {
                "description": "Internal Server Error",
                "schema": {
                  "$ref": "#/definitions/errors.AppError"
                }
              }
            }
          }
        },
      "/v1/deployments/{deploymentId}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍑 Deployment"
          ],
          "summary": "Get a specific deployment.",
          "parameters": [
            {
              "type": "integer",
              "description": "Deployment ID",
              "name": "deploymentId",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/deployments.DeploymentResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍑 Deployment"
          ],
          "summary": "Delete a specific deployment.",
          "parameters": [
            {
              "type": "integer",
              "description": "Deployment identifier",
              "name": "deploymentId",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            }
          }
        }
      },
      "/v1/deployments/{deploymentId}/activities": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                  "🍑 Deployment"
                ],
              "summary": "Get all activities of a specific deployment.",
                "parameters": [
                    {
                      "type": "integer",
                      "description": "Deployment ID",
                      "name": "deploymentId",
                      "in": "path",
                      "required": true
                    },
                  {
                    "minimum": 1,
                        "type": "integer",
                        "default": 1,
                    "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                  {
                    "type": "string",
                    "default": "id",
                    "description": "Order by",
                    "name": "order_by",
                    "in": "query"
                  },
                  {
                    "type": "string",
                    "default": "desc",
                    "description": "Order direction",
                    "name": "order_direction",
                    "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                              "$ref": "#/definitions/activities.ActivityResponse"
                            }
                        }
                    },
                  "404": {
                    "description": "Not Found",
                    "schema": {
                      "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
      },
      "/v1/deployments/{deploymentId}/module": {
        "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                  "🍑 Deployment"
                ],
          "summary": "Get the module of a specific deployment.",
                "parameters": [
                    {
                      "type": "integer",
                      "description": "Deployment ID",
                      "name": "deploymentId",
                      "in": "path",
                      "required": true
                    }
                ],
                "responses": {
                  "200": {
                    "description": "OK",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                  },
                  "404": {
                    "description": "Not Found",
                    "schema": {
                      "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
      "/v1/deployments/{deploymentId}/module/source": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                  "🍑 Deployment"
                ],
              "summary": "Get the source of the module of a specific deployment.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                      "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
      "activities.ActivityResponse": {
            "type": "object",
            "properties": {
              "createdAt": {
                    "type": "string"
                },
              "deploymentId": {
                "type": "integer"
              },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
              "type": {
                    "type": "string"
                }
            }
        },
      "deployment.DeploymentStatus": {
            "type": "string",
            "enum": [
                "unknown",
                "pending",
                "running",
                "failed",
                "succeeded"
            ],
            "x-enum-varnames": [
                "Unknown",
                "Pending",
                "Running",
                "Failed",
                "Succeeded"
            ]
        },
      "deployments.DeploymentRequest": {
            "type": "object",
            "properties": {
              "moduleID": {
                "type": "integer"
              },
                "name": {
                    "type": "string"
                },
              "variables": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/deployments.DeploymentVariableRequest"
                }
                }
            }
        },
      "deployments.DeploymentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
              "module_id": {
                "type": "integer"
                },
                "name": {
                    "type": "string"
                },
              "status": {
                "$ref": "#/definitions/deployment.DeploymentStatus"
                }
            }
        },
      "deployments.DeploymentVariableRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "deployments.ListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/deployments.DeploymentResponse"
            }
          }
        }
      },
      "errors.AppError": {
        "type": "object",
        "properties": {
          "err": {},
          "type": {
            "type": "string"
          }
        }
      },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}